Refactor to isolate websocket related code.
